requirements.txt
pytest~=8.4.1

requests~=2.32.4
python-dotenv~=1.1.1